refactor(graph): extract sigma global shared types and events
