+++ packages/graph-engine/src/render/sigma-events.ts
@@ -1,0 +1,14 @@
+export interface SigmaGlobalPointerEventPayload {
+  node?: unknown;
+  event?: { x?: unknown; y?: unknown; preventSigmaDefault?: () => void };
+  x?: unknown;
+  y?: unknown;
+  preventSigmaDefault?: () => void;
+}
+
+export function preventSigmaDefault(payload: unknown): void {
+  const eventPayload = payload as SigmaGlobalPointerEventPayload | null;
+  eventPayload?.preventSigmaDefault?.();
+  eventPayload?.event?.preventSigmaDefault?.();
+  if (payload instanceof Event) payload.preventDefault();
+}
